backend/requirements.txt
flask
flask-cors
pytesseract
pillow
pdfplumber
pdf2image
sentence-transformers
scikit-learn
networkx
